pyproject.toml
[project]
name = "zenn-ai-agent-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "google-cloud>=0.34.0",
    "google-cloud-storage>=2.19.0",
    "streamlit>=1.41.1",
    "uvicorn>=0.34.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.20",
    "google-cloud-aiplatform>=1.75.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
